==> LPS/analysis/2-annotation.Rmd <==
---
title: "Cluster annotation"
author: 
- name: Pierre-Luc Germain
  affiliation:
  - &IMLS Institute for Molecular Life Sciences, University of Zurich, Switzerland
- name: Helena L. Crowell
  affiliation: 
  - *IMLS
  - Swiss Institute of Bioinformatics (SIB), University of Zurich, Switzerland
date: '`r format(Sys.Date(), "%B %d, %Y")`'
output:
  html_document:
    toc: true
    code_folding: hide
---

```{r setup, include = FALSE}
knitr::opts_knit$set(root.dir = "..")
knitr::opts_chunk$set(autodep = TRUE, cache = TRUE, cache.lazy = FALSE)
```

### Load packages

```{r load-libs, message = FALSE, warning = FALSE}
library(ComplexHeatmap)
library(cowplot)
library(ggplot2)
library(dplyr)
library(muscat)
library(purrr)
library(RColorBrewer)
library(viridis)
library(scran)
library(Seurat)
library(SingleCellExperiment)
```

# Load data & convert to SCE

```{r load-data}
so <- readRDS(file.path("output", "SO_clustering.rds"))
sce <- as.SingleCellExperiment(so, assay = "RNA")
colData(sce) <- as.data.frame(colData(sce)) %>% 
    mutate_if(is.character, as.factor) %>% 
    DataFrame(row.names = colnames(sce))
```

# Nb. of clusters by resolution

```{r}
cluster_cols <- grep("res.[0-9]", colnames(colData(sce)), value = TRUE)
sapply(colData(sce)[cluster_cols], nlevels)
```

# Cluster-sample counts

```{r}
# set cluster IDs to resolution 0.2 clustering
so <- SetIdent(so, value = "integrated_snn_res.0.2")

so@meta.data$cluster_id <- Idents(so)
sce$cluster_id <- Idents(so)
(n_cells <- table(sce$cluster_id, sce$sample_id))
```

# Relative cluster-abundances

```{r fig.width = 6, fig.height = 5}
fqs <- prop.table(n_cells, margin = 2)
mat <- as.matrix(unclass(fqs))
Heatmap(mat,
    col = rev(brewer.pal(11, "RdGy")[-6]),
    name = "Frequency",
    cluster_rows = FALSE,
    cluster_columns = FALSE,
    row_names_side = "left",
    row_title = "cluster_id",
    column_title = "sample_id",
    column_title_side = "bottom",
    rect_gp = gpar(col = "white"),
    cell_fun = function(i, j, x, y, width, height, fill)
        grid.text(round(mat[j, i] * 100, 2), x = x, y = y, 
            gp = gpar(col = "white", fontsize = 8)))
```

# DR colored by cluster ID {.tabset}

```{r dim-red, results = "asis", fig.width = 12}
cs <- sample(colnames(so), 5e3)
.plot_dr <- function(so, dr, id)
    DimPlot(so, cells = cs, group.by = id, reduction = dr, pt.size = 0.4) +
        guides(col = guide_legend(nrow = 11, 
            override.aes = list(size = 3, alpha = 1))) +
        theme_void() + theme(aspect.ratio = 1)

ids <- c("cluster_id", "group_id", "sample_id")
for (id in ids) {
    cat("## ", id, "\n")
    p1 <- .plot_dr(so, "tsne", id)
    lgd <- get_legend(p1)
    p1 <- p1 + theme(legend.position = "none")
    p2 <- .plot_dr(so, "umap", id) + theme(legend.position = "none")
    ps <- plot_grid(plotlist = list(p1, p2), nrow = 1)
    p <- plot_grid(ps, lgd, nrow = 1, rel_widths = c(1, 0.2))
    print(p)
    cat("\n\n")
}
```

# Known marker genes

```{r known-markers}
fs <- list(
    astrocytes = c("Aqp4", "Gfap", "Fgfr3"),
    endothelial = c("Cldn5","Nostrin"),
    microglia = c("C1qb","Tyrobp"),
    neuronal = c("Snap25", "Stmn2"),
    neuronal_excitatory = "Slc17a7",
    neuronal_inhibitory = "Gad1",
    oligodendrocyte = "Opalin",
    OPC = "Pdgfra", CPE = "Ttr")

fs <- lapply(fs, sapply, function(g) 
    grep(paste0(g, "$"), rownames(sce), value = TRUE))

gs <- gsub(".*\\.", "", unlist(fs))
ns <- vapply(fs, length, numeric(1))
ks <- rep.int(names(fs), ns)
labs <- sprintf("%s(%s)", gs, ks)
```

## Heatmap of mean marker-exprs. by cluster

```{r known-markers-heatmap, fig.width = 8, fig.height = 4}
# split cells by cluster
cs_by_k <- split(colnames(sce), sce$cluster_id)

# compute cluster-marker means
ms_by_cluster <- lapply(fs, function(gs)
    vapply(cs_by_k, function(i)
        Matrix::rowMeans(logcounts(sce)[gs, i, drop = FALSE]), 
        numeric(length(gs))))

# prep. for plotting & scale b/w 0 and 1
mat <- do.call("rbind", ms_by_cluster)
mat <- muscat:::.scale(mat)
rownames(mat) <- gs

cols <- muscat:::.cluster_colors[seq_along(fs)]
cols <- setNames(cols, names(fs))
row_anno <- rowAnnotation(
    df = data.frame(label = factor(ks, levels = names(fs))),
    col = list(label = cols), gp = gpar(col = "white")) 
Heatmap(mat,
    name = "scaled avg.\nexpression",
    col = viridis(10),
    cluster_rows = FALSE,
    cluster_columns = FALSE,
    row_names_side = "left",
    column_title = "cluster_id",
    column_title_side = "bottom",
    rect_gp = gpar(col = "white"),
    left_annotation = row_anno)
```

## DR colored by marker expr.

```{r dim-red-expr, fig.width = 12, fig.height = 12}
# downsample to 2000 cells
cs <- sample(colnames(sce), 2e3)
sub <- subset(so, cells = cs)
gs <- unlist(fs)

# UMAPs colored by marker-expression
ps <- lapply(seq_along(gs), function(i) {
    if (!gs[i] %in% rownames(so)) return(NULL)
    FeaturePlot(sub, features = gs[i], reduction = "umap", pt.size = 0.4) +
        theme(aspect.ratio = 1, legend.position = "none") +
        ggtitle(labs[i]) + theme_void()
})

# arrange plots in grid
ps <- ps[!vapply(ps, is.null, logical(1))]
plot_grid(plotlist = ps, ncol = 4, label_size = 10)
```

# Find markers using `scran`

```{r scran-findMarkers}
scran_markers <- findMarkers(sce, 
    groups = sce$cluster_id, block = sce$sample_id, 
    direction = "up", lfc = 2, full.stats = TRUE)
```

## Heatmap of mean marker-exprs. by cluster

```{r scran-markers-heatmap, fig.width = 12, fig.height = 6}
gs <- lapply(scran_markers, function(u) rownames(u)[u$Top == 1])
sub <- sce[unique(unlist(gs)), ]
pbs <- aggregateData(sub, assay = "logcounts", by = "cluster_id", fun = "mean")
mat <- t(muscat:::.scale(assay(pbs)))

Heatmap(mat,
    name = "scaled avg.\nexpression",
    col = viridis(10),
    cluster_rows = FALSE,
    cluster_columns = FALSE,
    row_names_side = "left",
    row_title = "cluster_id",
    rect_gp = gpar(col = "white"))
```

## Cluster annotation

In consideration of the above visualizations and additional exploration with `r BiocStyle::Biocpkg("iSEE")`, we arrive at the following cluster annotations: 

```{r cluster-anno, warning = FALSE}
anno <- list(
    "unassigned" = 18,
    "Astrocytes" = 3,
    "Endothelial" = 13,
    "Microglia" = c(19, 21),
    "Oligodendrocytes" = 4,
    "OPC" = 14,
    "CPE cells" = 20,
    "Excit. Neuron" = c(0, 1, 2, 5, 6, 7, 8, 11, 15, 17),
    "Inhib. Neuron" = c(9, 10, 12, 16))

m <- match(sce$cluster_id, unlist(anno))
ns <- vapply(anno, length, numeric(1))
lab <- rep.int(names(anno), ns)[m]
sce$cluster_id <- factor(lab, levels = names(anno)[-1])
sce <- sce[, !is.na(sce$cluster_id)]

# cluster-sample cell-counts
table(sce$cluster_id, sce$sample_id)
```

# Save SCE to .rds

```{r save-data}
# normalize for visualization
sce <- logNormCounts(sce)

# make WT reference group & rename
sce$group_id <- factor(sce$group_id, 
  levels = c("WT", "LPS"), 
  labels = c("Vehicle", "LPS"))

# reorder sample levels
m <- match(levels(sce$sample_id), sce$sample_id)
o <- order(sce$group_id[m])
sce$sample_id <- factor(sce$sample_id, 
    levels = levels(sce$sample_id)[o])

# separate ensembl IDs & gene symbols
ss <- strsplit(rownames(sce), ".", fixed=TRUE)
rowData(sce)$ensembl_id <- sapply(ss, .subset, 1)
rowData(sce)$symbol <- sapply(ss, .subset, 2)

# prep. SCE for `muscat` & write to .rds
saveRDS(prepSCE(sce), file.path("output", "SCE_annotation.rds"))
```
