==> LPS/analysis/5-geneset.Rmd <==
---
title: "Geneset analysis"
author: 
- name: Mark D. Robinson
  affiliation: 
  - &IMLS Institute for Molecular Life Sciences, University of Zurich, Switzerland
  - &SIB Swiss Institute of Bioinformatics (SIB), University of Zurich, Switzerland
- name: Helena L. Crowell
  affiliation: 
  - *IMLS
  - *SIB
date: '`r format(Sys.Date(), "%B %d, %Y")`'
output:
  html_document:
    toc: true
    code_folding: show
---

```{r setup, include = FALSE}
knitr::opts_knit$set(root.dir = "..")
knitr::opts_chunk$set(autodep = TRUE, cache = TRUE, cache.lazy = FALSE)
```

```{r load-libs, warning = FALSE, message = FALSE}
library(dplyr)
library(edgeR)
library(limma)
library(magrittr)
library(msigdbr)
library(muscat)
library(purrr)
library(dplyr)
library(reshape2)
library(pheatmap)
```

### Load data

```{r load-data}
# load data & logFCs by cluster-sample
sce <- readRDS(file.path("output", "SCE_annotation.rds"))
lfc <- readRDS(file.path("output", "logFCs_by_cluster-sample.rds"))

# source utilities
source(file.path("code", "utils.R"))
```

# Pseudobulk DS analysis

```{r ds-analysis}
# compute pseudobulks (sum of counts) by cluster-sample & run DS analysis
pbs <- aggregateData(sce)
res <- pbDS(pbs, method = "edgeR", filter = "genes", verbose = FALSE)
```

# Organize genesets & names

```{r prep-data}
# separate ensembl ID & gene symbols
dat <- lapply(res$data, function(u) {
  ss <- strsplit(rownames(u), ".", fixed=TRUE)
  u$genes <- data.frame(
      ensembl_id = sapply(ss, .subset, 1), 
      symbol = sapply(ss, .subset, 2))
  return(u)
})

m_df <- msigdbr(species = "Mus musculus")
m_df <- filter(m_df, gs_cat %in% c("H", "C5", "C7"))

sapply(dat, function(u) table(u$genes$symbol %in% m_df$gene_symbol))
sets <- split(m_df$gene_symbol, m_df$gs_name)
n <- vapply(sets, length, numeric(1))
sets <- sets[n >= 20 & n <= 1000]
length(sets)
```

# Run `camera` on ea. cluster

```{r run-camera}
gs_dat <- mapply(function(uu, vv) {
    inds <- ids2indices(sets, uu$genes$symbol, remove.empty = TRUE)
    mm <- uu$design[colnames(uu),]
    v <- voom(uu, mm)
    f <- lmFit(v, mm)
    f <- eBayes(f)
    c <- which(colnames(mm) == "LPS")
    cf <- contrasts.fit(f, coefficients = c)
    cf <- eBayes(cf)
    list(indices = inds, voom = v, design = mm, 
        cluster_id = vv, contrasts.fit = cf)
}, dat, names(dat), SIMPLIFY = FALSE)

gs_df <- lapply(gs_dat, function(u)
    camera(u$voom, u$indices, u$design) %>% 
        mutate(geneset = rownames(.))) %>% 
    bind_rows(.id = "cluster_id")
```

# Heatmap summary

```{r heatmap-summary, fig.width = 14, fig.height = 10}
length(cats <- unique(pull(filter(gs_df, FDR < 1e-10), geneset)))
filter(gs_df, geneset %in% cats) %>%
    mutate(neg_log10_fdr = -log10(FDR)) %>% 
    acast(cluster_id ~ geneset, value.var = "neg_log10_fdr") %>% 
    set_colnames(gsub("/*([^_]*)_(.*)", "\\2", colnames(.))) %>% 
    set_colnames(strtrim(colnames(.), 30)) %>%
    pheatmap(fontsize = 8, border_color = NA, color = colorRampPalette(
        c("aliceblue", "cornflowerblue", "violet", "red"))(50))
```

# logFCs by cluster-sample

## GO_CHEMOKINE_RECEPTOR_BINDING

```{r lfc-heatmap-chemo, fig.width = 14, fig.height = 8}
gs <- unique(sets[["GO_CHEMOKINE_RECEPTOR_BINDING"]])
m <- match(gs, rowData(lfc)$symbol, nomatch = 0)
qs <- quantile(assay(lfc[m, ]), c(0.01, 0.99))
brks <- .get_brks(qs, 0.5)
cols <- colorRamp2(brks, .hm_cols)

col_df <- data.frame(
    cluster_id = rep(levels(lfc$cluster_id), each = ns), 
    group_id = rep(pbs$group_id, nlevels(lfc$cluster_id)))
col_cols <- sapply(ids, function(id) get(paste0(id, "_pal")))
col_anno <- HeatmapAnnotation(
    df = col_df, col = col_cols,
    show_annotation_name = FALSE,
    annotation_legend_param = .hm_lgd_aes)

tmp <- lfc[m, ]
rownames(tmp) <- gsub(".*[.]{1}", "", rownames(tmp))
(hm <- .plot_hm(tmp, cols, brks, 
    col_anno, row_title = NULL, row_nms = TRUE,
    cluster_row_slices = FALSE, cluster_column_slices = FALSE, 
    row_names_gp = gpar(fontsize = 5),
    column_title_gp = gpar(fontsize = 8),
    use_raster = TRUE, raster_device = "CairoPNG"))
```

```{r echo = FALSE}
ggsave(file.path("figures", "lfc_heatmap_chemo.pdf"), 
    grid.draw(grid.grabExpr(draw(hm))),
    width = 15, height = 7, units = "cm",
    dpi = 300, useDingbats = FALSE)
```

## IFN-alpha XOR -gamma

```{r lfc-heatmap-ifnag}
igs <- sapply(c("interferon_alpha", "interferon_gamma"), 
    grep, names(sets), ignore.case = TRUE) %>% 
    sapply(function(u) unlist(sets[u]))
gs <- setdiff(unlist(igs), intersect(igs[[1]], igs[[2]]))

# subset unique DS genes
ds_gs <- .get_ds_gs(res$table$LPS, fdr = 0.05, lfc = 1, simplify = TRUE)
m <- match(gs, rowData(lfc[ds_gs, ])$symbol, nomatch = 0)

qs <- quantile(assay(lfc[m, ]), c(0.01, 0.99))
brks <- .get_brks(qs, 1)
cols <- colorRamp2(brks, .hm_cols)

# split rows by alpha/gamma
w <- as.numeric(gs[m != 0] %in% igs[[1]]) + 1
row_cols <- c("skyblue", "lightgreen")
names(row_cols) <- c("IFN-alpha", "IFN-beta")

row_anno <- rowAnnotation(
    show_annotation_name = FALSE,
    df = data.frame(response = names(row_cols)[w]),
    col = list(response = row_cols),
    simple_anno_size = unit(1, "mm"),
    annotation_legend_param = .hm_lgd_aes)

tmp <- lfc[m, ]
rownames(tmp) <- gsub(".*[.]{1}", "", rownames(tmp))
(hm <- .plot_hm(tmp, cols, brks, 
    col_anno, row_anno, row_split = w, row_title = NULL, row_nms = TRUE,
    cluster_row_slices = FALSE, cluster_column_slices = FALSE, 
    row_names_gp = gpar(fontsize = 5),
    column_title_gp = gpar(fontsize = 8),
    use_raster = TRUE, raster_device = "CairoPNG"))
```

```{r echo = FALSE}
ggsave(file.path("figures", "lfc_heatmap_ifnag.pdf"), 
    grid.draw(grid.grabExpr(draw(hm))),
    width = 15, height = 12, units = "cm",
    dpi = 300, useDingbats = FALSE)
```

# Barcode plots {.tabset}

```{r barcode-plots, results = "asis", fig.width = 6, fig.height = 4}
cats_by_k <- gs_df %>% 
    group_by(cluster_id) %>% 
    top_n(10, dplyr::desc(FDR)) %>% 
    group_split %>% 
    lapply(pull, geneset) %>% 
    set_names(names(res$data))

lapply(names(res$data), function(k) {
    cat("## ", k, "\n")
    lapply(cats_by_k[[k]], function(c)
        barcodeplot(
            statistics = gs_dat[[k]]$contrasts.fit$t[, 1],
            index = gs_dat[[k]]$indices[[c]],
            quantiles=c(-1, 1) * qt(0.95, df = 14),
            main = c, cex.main = 0.8))
    cat("\n\n")
})
```

# Write results to .rds & .csv

```{r save-res}
saveRDS(gs_dat, file.path("output", "GSEA_data.rds")) 
saveRDS(gs_df, file.path("output", "GSEA_df.rds")) 
write.csv(dplyr::filter(gs_df, FDR < 0.05),
    file.path("output", "GSEA_results.csv"),
    quote = FALSE, row.names = FALSE)
```
